[doc] Highlight making the application public
A number of operator authors fail to notice that unless they provide a
secret the application needs to be public.

After reviewing the testing-opreator.md document, its difficult to
notice the comment about making the application public because it is
included in a quote. This change moves the comment from a quote to
regular text.